move CreatePost page to Dashboard
