+++ src/components/view/dashboard/create-post/form/handle-create-post/handleCreatePost.js
@@ -1,0 +1,16 @@
+import Api from "../../../../../../service/gateway/Api";
+
+export const fetchTags = async (setTags) => {
+  try {
+    const apiInstance = Api.getInstance();
+    const response = await apiInstance.getTagAll();
+    if (response) {
+      setTags(response.tags);
+    } else {
+      setTags([]);
+    }
+  } catch (error) {
+    console.error("Erro ao buscar tags:", error);
+    setTags([]);
+  }
+};
